refactor: promptmessage 클래스의 위치변경, outputview 클래스 추가
