+++ src/main/java/lotto/custom/controller/LottoController.java
@@ -2,13 +2,16 @@ package lotto.custom.controller;
 
 import lotto.custom.service.LottoPurchaseService;
 import lotto.custom.view.InputView;
+import lotto.custom.view.OutputView;
 
 public class LottoController {
     private final InputView inputView;
+    private final OutputView outputView;
     private final LottoPurchaseService lottoPurchaseService;
 
     public LottoController() {
         this.inputView = new InputView();
+        this.outputView = new OutputView();
         this.lottoPurchaseService = new LottoPurchaseService();
     }
 
@@ -19,7 +22,7 @@ public class LottoController {
                 lottoPurchaseService.run(purchaseAmountInput);
                 break; // 예외가 발생하지 않으면 루프 종료
             } catch (IllegalArgumentException e) {
-                System.out.println(e.getMessage());
+                outputView.displayErrorMessage(e.getMessage());
             }
         }
     }
