+++ src/main/java/lotto/custom/view/OutputView.java
@@ -1,0 +1,7 @@
+package lotto.custom.view;
+
+public class OutputView {
+    public void displayErrorMessage(String errorMessage) {
+        System.out.println(errorMessage);
+    }
+}
