+++ src/main/java/lotto/custom/view/PromptMessages.java
@@ -1,4 +1,4 @@
-package lotto.custom.constants;
+package lotto.custom.view;
 
 public class PromptMessages {
     public static final String INPUT_PURCHASE_AMOUNT_PROMPT = "구입금액을 입력해 주세요.";
